Mccalluc/v0.0.21 (#366)
* Version bump

* tidy changelog

* Update CHANGELOG

* set version var

* tweak push again

* Only update changelog if needed

* Remove stray file

+++ push.sh
@@ -8,18 +8,22 @@ git diff --quiet || die 'Uncommitted changes: Stash or commit before push.'
 diff VERSION <(curl -s https://raw.githubusercontent.com/hubmapconsortium/portal-ui/master/context/app/static/VERSION) \
   && die 'Update VERSION and commit.'
 
-(
-  echo '##' `cat VERSION` - `date +"%F"`
-  echo
-  cat CHANGELOG-*.md
-  echo
-  cat CHANGELOG.md
-) > CHANGELOG.md.new
-mv CHANGELOG.md.new CHANGELOG.md
-git rm CHANGELOG-*.md
-git add .
-git commit -m 'Update CHANGELOG'
+if ls CHANGELOG-*.md; then
+  (
+    echo '##' `cat VERSION` - `date +"%F"`
+    echo
+    cat CHANGELOG-*.md
+    echo
+    echo
+    cat CHANGELOG.md
+  ) > CHANGELOG.md.new
+  mv CHANGELOG.md.new CHANGELOG.md
+  git rm CHANGELOG-*.md
+  git add .
+  git commit -m 'Update CHANGELOG'
+fi
 
+VERSION=`cat VERSION`
 VERSION_IMAGE_NAME=hubmap/portal-ui:$VERSION
 LATEST_IMAGE_NAME=hubmap/portal-ui:latest
